+++ app/controllers/v1/router/CampaignApiRouter.scala
@@ -1,7 +1,7 @@
-package controllers.controllerRoutes
+package controllers.v1.router
 
-import controllers.apis.CampaignsApiController
 import controllers.controllerRoutes.routerGeneric.AbstractGenericRouter
+import controllers.v1.apis.CampaignsApiController
 import javax.inject.Inject
 import shared.com.ortb.persistent.schema.Tables._
